refactor(cli): share strict fixture target selection
