+++ crates/openlogi-cli/src/cmd/fixture/record_profile/selection.rs
@@ -1,10 +1,10 @@
-//! Strict physical target selection with privacy-safe diagnostics.
+//! Physical target projection for semantic profile capture.
 
-use std::fmt::Write as _;
-
-use anyhow::{Result, anyhow, bail};
+use anyhow::Result;
 use openlogi_core::device::{DeviceInventory, StandaloneDevice};
 use openlogi_core::hid::DeviceRoute;
+
+use super::super::target_selection::{self, FixtureTarget};
 
 #[derive(Clone, Copy, Debug, PartialEq, Eq)]
 pub(super) enum TargetLocation {
@@ -17,6 +17,16 @@ pub(super) struct TargetCandidate {
     route: DeviceRoute,
     name: String,
     pub(super) location: TargetLocation,
+}
+
+impl FixtureTarget for TargetCandidate {
+    fn route(&self) -> &DeviceRoute {
+        &self.route
+    }
+
+    fn display_name(&self) -> &str {
+        &self.name
+    }
 }
 
 pub(super) fn target_candidates(
@@ -61,76 +71,7 @@ pub(super) fn select_target(
     candidates: &[TargetCandidate],
     query: Option<&str>,
 ) -> Result<TargetCandidate> {
-    if candidates.is_empty() {
-        bail!("the Agent snapshot contains no addressable physical device candidate");
-    }
-
-    let mut matches = match query {
-        Some(query) => candidates
-            .iter()
-            .filter(|candidate| {
-                candidate.name.eq_ignore_ascii_case(query) || candidate.route.to_string() == query
-            })
-            .cloned()
-            .collect::<Vec<_>>(),
-        None if candidates.len() == 1 => candidates.to_owned(),
-        None => Vec::new(),
-    };
-    if matches.len() != 1 {
-        let message = match (query, matches.len()) {
-            (Some(_), 0) => "no candidate exactly matched --device",
-            (Some(_), _) => "--device matched more than one candidate",
-            (None, _) => "more than one physical candidate is available; pass --device",
-        };
-        return Err(selection_error(message, candidates));
-    }
-
-    let selected = matches.remove(0);
-    if candidates
-        .iter()
-        .filter(|candidate| candidate.route == selected.route)
-        .count()
-        != 1
-    {
-        return Err(selection_error(
-            "the selected route is duplicated and cannot identify one target",
-            candidates,
-        ));
-    }
-    Ok(selected)
-}
-
-fn selection_error(message: &str, candidates: &[TargetCandidate]) -> anyhow::Error {
-    let mut list = String::new();
-    for candidate in candidates {
-        let _ = write!(
-            list,
-            "\n  - {:?} ({})",
-            candidate.name,
-            safe_route_label(&candidate.route)
-        );
-    }
-    anyhow!("{message}. Candidates:{list}")
-}
-
-fn safe_route_label(route: &DeviceRoute) -> String {
-    match route {
-        DeviceRoute::Bolt { slot, .. } => format!("Bolt receiver slot {slot}"),
-        DeviceRoute::Unifying { slot, .. } => format!("Unifying receiver slot {slot}"),
-        DeviceRoute::Direct {
-            vendor_id,
-            product_id,
-        } => format!("direct {vendor_id:04x}:{product_id:04x}"),
-        DeviceRoute::RawHid {
-            vendor_id,
-            product_id,
-            usage_page,
-            usage_id,
-            ..
-        } => format!(
-            "standalone {vendor_id:04x}:{product_id:04x} usage {usage_page:04x}:{usage_id:04x}"
-        ),
-    }
+    target_selection::select_target(candidates, query)
 }
 
 pub(super) fn standalone_route(device: &StandaloneDevice) -> DeviceRoute {
